refactor: move chunk hydration status to storage
