+++ src/Arius.Core.Tests/ChunkStorage/ChunkStorageServiceReadTests.cs
@@ -1,4 +1,3 @@
-using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.ChunkStorage;
 using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.Storage;
