+++ src/Arius.Core.Tests/List/ResolveFileHydrationStatusesHandlerTests.cs
@@ -1,12 +1,13 @@
 using Arius.Core.Features.ChunkHydrationStatusQuery;
 using Arius.Core.Features.ListQuery;
-using Arius.Core.Tests.Fakes;
 using Arius.Core.Shared.ChunkIndex;
+using Arius.Core.Tests.Fakes;
 using Arius.Core.Shared.ChunkStorage;
 using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.Storage;
 using Microsoft.Extensions.Logging.Abstractions;
 using Shouldly;
+using ChunkHydrationStatus = Arius.Core.Shared.ChunkStorage.ChunkHydrationStatus;
 
 namespace Arius.Core.Tests.List;
