+++ src/Arius.Core/Shared/ChunkStorage/IChunkStorageService.cs
@@ -1,4 +1,3 @@
-using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.Storage;
 
 namespace Arius.Core.Shared.ChunkStorage;
