+++ src/Arius.Explorer/RepositoryExplorer/FileItemViewModel.cs
@@ -1,5 +1,5 @@
 using Arius.Core.Features.ListQuery;
-using Arius.Core.Shared.ChunkIndex;
+using Arius.Core.Shared.ChunkStorage;
 using CommunityToolkit.Mvvm.ComponentModel;
 using System.IO;
 using System.Windows.Media;
